Add search widget (WIP)
